feat: use hono vite plugins
