+++ eslint.config.js
@@ -7,7 +7,7 @@ import reactRefresh from 'eslint-plugin-react-refresh';
 /** @type {import("eslint").Linter.FlatConfig[]} */
 export default [
   {
-    ignores: ['**/node_modules'],
+    ignores: ['**/node_modules', 'dist'],
   },
   js.configs.recommended,
   {
