--- functions/api/[[route]].ts
@@ -1,4 +1,0 @@
-import { handle } from 'hono/cloudflare-pages';
-import { app } from '../../src/api';
-
-export const onRequest = handle(app);
