+++ vite.config.ts
@@ -1,10 +1,43 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react-swc';
+import pages from '@hono/vite-cloudflare-pages';
+import devServer from '@hono/vite-dev-server';
 
-// https://vitejs.dev/config/
-export default defineConfig({
+const globalConfig = {
   resolve: {
     alias: [{ find: '@', replacement: '/src' }],
   },
-  plugins: [react()],
+};
+
+export default defineConfig(({ mode }) => {
+  if (mode === 'client') {
+    return {
+      ...globalConfig,
+      build: {
+        rollupOptions: {
+          input: ['./src/client.tsx', './src/index.css'],
+          output: {
+            entryFileNames: 'static/[name].js',
+            assetFileNames: 'static/[name].[ext]',
+          },
+        },
+      },
+    };
+  } else if (mode === 'test') {
+    return {
+      ...globalConfig,
+    };
+  } else {
+    return {
+      ...globalConfig,
+      ssr: {
+        external: ['react', 'react-dom'],
+      },
+      plugins: [
+        pages(),
+        devServer({
+          entry: 'src/index.tsx',
+        }),
+      ],
+    };
+  }
 });
